ACF 5: "get_as_timestamp" handling.
If we have saved the value as a timestamp, `format_value()` converts it to a formatted value iff `get_as_timestamp` isn't true.

If we have saved the value as a timestamp, `render_field()` always converts it to a formatted value for the date/time pickers.

Removed the unnecessary `load_value()`.

Introduced private helper functions in order to remain DRY.